backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
motor>=3.0.0
pymongo>=4.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
google-generativeai>=0.4.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
aiofiles>=22.0.0
httpx>=0.23.0
joblib>=1.3.0
pandas>=2.0.0
scikit-learn>=1.3.0
openpyxl>=3.1.0
numpy>=1.24.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
